Rename file and Cleanup. Remove excluded types.
