--- ios/brave-ios/Sources/Brave/Frontend/Share/TitleActivityItemProvider.swift
@@ -1,103 +1,0 @@
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this
-// file, You can obtain one at https://mozilla.org/MPL/2.0/.
-
-import Foundation
-import UIKit
-import UniformTypeIdentifiers
-
-/// This Activity Item Provider subclass does two things that are non-standard behaviour:
-///
-/// * We return NSNull if the calling activity is not supposed to see the title. For example the Copy action, which should only paste the URL. We also include Message and Mail to have parity with what Safari exposes.
-/// * We set the subject of the item to the title, this means it will correctly be used when sharing to for example Mail. Again parity with Safari.
-///
-/// Note that not all applications use the Subject. For example OmniFocus ignores it, so we need to do both.
-
-class TitleActivityItemProvider: NSObject, UIActivityItemSource {
-  let title: String
-
-  static let activityTypesToIgnore = [
-    UIActivity.ActivityType.copyToPasteboard, UIActivity.ActivityType.message,
-    UIActivity.ActivityType.mail,
-  ]
-
-  init(title: String) {
-    self.title = title
-    super.init()
-  }
-
-  func activityViewControllerPlaceholderItem(
-    _ activityViewController: UIActivityViewController
-  ) -> Any {
-    return title
-  }
-
-  func activityViewController(
-    _ activityViewController: UIActivityViewController,
-    itemForActivityType activityType: UIActivity.ActivityType?
-  ) -> Any? {
-    if let activityType = activityType {
-      if TitleActivityItemProvider.activityTypesToIgnore.contains(activityType) {
-        return nil
-      }
-    }
-    return title
-  }
-
-  func activityViewController(
-    _ activityViewController: UIActivityViewController,
-    subjectForActivityType activityType: UIActivity.ActivityType?
-  ) -> String {
-    return title
-  }
-
-  func activityViewController(
-    _ activityViewController: UIActivityViewController,
-    dataTypeIdentifierForActivityType activityType: UIActivity.ActivityType?
-  ) -> String {
-    return UTType.text.identifier
-  }
-}
-
-class URLActivityItemProvider: NSObject, UIActivityItemSource {
-  let title: String
-  let url: URL
-
-  init(title: String, url: URL) {
-    self.title = title
-    self.url = url
-    super.init()
-  }
-
-  func activityViewControllerPlaceholderItem(
-    _ activityViewController: UIActivityViewController
-  ) -> Any {
-    return url
-  }
-
-  func activityViewController(
-    _ activityViewController: UIActivityViewController,
-    itemForActivityType activityType: UIActivity.ActivityType?
-  ) -> Any? {
-    if let activityType = activityType {
-      if TitleActivityItemProvider.activityTypesToIgnore.contains(activityType) {
-        return nil
-      }
-    }
-    return url
-  }
-
-  func activityViewController(
-    _ activityViewController: UIActivityViewController,
-    subjectForActivityType activityType: UIActivity.ActivityType?
-  ) -> String {
-    return title
-  }
-
-  func activityViewController(
-    _ activityViewController: UIActivityViewController,
-    dataTypeIdentifierForActivityType activityType: UIActivity.ActivityType?
-  ) -> String {
-    return UTType.url.identifier
-  }
-}
