+++ ios/brave-ios/Sources/Brave/Frontend/Share/URLActivityItemProvider.swift
@@ -1,0 +1,45 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+import Foundation
+import UIKit
+import UniformTypeIdentifiers
+
+class URLActivityItemProvider: NSObject, UIActivityItemSource {
+  let title: String
+  let url: URL
+
+  init(title: String, url: URL) {
+    self.title = title
+    self.url = url
+    super.init()
+  }
+
+  func activityViewControllerPlaceholderItem(
+    _ activityViewController: UIActivityViewController
+  ) -> Any {
+    return url
+  }
+
+  func activityViewController(
+    _ activityViewController: UIActivityViewController,
+    itemForActivityType activityType: UIActivity.ActivityType?
+  ) -> Any? {
+    return url
+  }
+
+  func activityViewController(
+    _ activityViewController: UIActivityViewController,
+    subjectForActivityType activityType: UIActivity.ActivityType?
+  ) -> String {
+    return title
+  }
+
+  func activityViewController(
+    _ activityViewController: UIActivityViewController,
+    dataTypeIdentifierForActivityType activityType: UIActivity.ActivityType?
+  ) -> String {
+    return UTType.url.identifier
+  }
+}
